requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
